+++ lib/blocs/habits/habits_cubit.dart
@@ -40,6 +40,8 @@ class HabitsCubit extends Cubit<HabitsState> {
             zeroBased: false,
             minY: 0,
             displayFilter: HabitDisplayFilter.openNow,
+            showSearch: false,
+            showTimeSpan: false,
           ),
         ) {
     _definitionsStream = _journalDb.watchHabitDefinitions();
@@ -222,6 +224,8 @@ class HabitsCubit extends Cubit<HabitsState> {
   var _shortStreakCount = 0;
   var _longStreakCount = 0;
   var _zeroBased = true;
+  var _showTimeSpan = false;
+  var _showSearch = false;
   var _timeSpanDays = isDesktop ? 14 : 7;
   var _selectedInfoYmd = '';
 
@@ -237,6 +241,16 @@ class HabitsCubit extends Cubit<HabitsState> {
 
   void toggleZeroBased() {
     _zeroBased = !_zeroBased;
+    emitState();
+  }
+
+  void toggleShowSearch() {
+    _showSearch = !_showSearch;
+    emitState();
+  }
+
+  void toggleShowTimeSpan() {
+    _showTimeSpan = !_showTimeSpan;
     emitState();
   }
 
@@ -295,6 +309,8 @@ class HabitsCubit extends Cubit<HabitsState> {
           state: state,
         ),
         displayFilter: _displayFilter,
+        showTimeSpan: _showTimeSpan,
+        showSearch: _showSearch,
       ),
     );
   }